// client/src/app/core/models/user.model.ts
export enum UserRole {
  PLAYER = 'player',
  COORDINATOR = 'coordinator',
}

export interface User {
  _id?: string;
  id?: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber?: string; // Added phone number field
  city?: string; // Added city field
  nickname?: string; // Added nickname field
  displayOnlyNickname?: boolean; // Added display preference for nickname
  auth0Id: string;
  role: UserRole;
  roles?: string[]; // Adding roles array for backwards compatibility
  isActive: boolean;
  skillLevel: number;
  preferSingles: boolean;
  preferDoubles: boolean; // Added preference for doubles
  photoUrl?: string; // Optional profile photo URL
  createdAt?: Date;
  updatedAt?: Date;
}